# swagger.yaml
swagger: '2.0'
info:
  title: Menu Maker API documentation
  description: Contains all available API endpoints in this codebase
  version: '1.0.0'
  termsOfService: 'http://swagger.io/terms/'
host: localhost:3001
basePath: /api/v1
schemes:
  - http
paths:
  /auth/login:
    post:
      tags:
        - Auth Module
      summary: Send the email for the connection
      description: API for Login
      parameters:
        - in: body
          name: body
          description: Login Payload
          required: true
          schema:
            $ref: '#/definitions/Login'
      produces:
        - application/json
      responses:
        '200':
          description: Email sent
        '400':
          description: Invalid Fields
        '500':
          description: Internal Server Error
  /auth/login/{token}:
    get:
      tags:
      - Auth Module
      summary: Get the auth token from email sent
      description: API for Login
      parameters:
        - in: path
          name: token
          type: string
          description: Short duration token
          required: true
      responses:
        "301":
          description: successfully authenticate
          headers:
            Credentials authToken:
              description: Auth token in httpOnly
              type: string
            Credentials isConnected:
              description: Value = 1 
              type: string
        "500":
          description: Internal Server Error
  /auth/logout:
    delete:
      tags:
      - Auth Module
      summary: Disconnect the user
      description: Disconnect the user by deleting the credential and invalidate the authToken
      parameters:
        - in: header
          type: string
          name: credentials
          description: send authToken credentials
      responses:
        "200":
          description: Successfully disconnected
          headers:
            Credentials authToken:
              description: set the expiration date at 0
              type: string
            Credentials isConnected:
              description: set the expiration date at 0
              type: string
        "500":
          description: Internal Server Error
  /menus/userMenus:
    get:
      security:
        - Bearer: []
      tags:
      - Menus Module
      summary: Get all menus created by user
      description: Retrieved the menus created by the user
      responses:
        "200":
          description: Menus retrieved successfully
          schema:
            $ref: "#/definitions/UserMenusResponse"
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
  /menus/createMenu:
    post:
      security:
       - Bearer: []
      tags:
      - Menus Module
      summary: Create a menu
      description: Create a menu from the user
      parameters:
        - in: body
          name: menu informations
          schema: 
            $ref: "#/definitions/MenuData"
          description: object with menu data
          required: true
      responses:
        "200":
          description: Menu created successfully
        "400":
          description: Bad request in body
        "401":
          description: Unauthorized
        "500":
          description: Interal Server Error
  /menus/{menuId}:
    get:
      security:
       - Bearer: []
      tags:
      - Menus Module
      summary: Get a specified menu
      description: Get a menu from the user with the menu id
      parameters:
        - in: path
          type: string
          name: menuId
          description: id of the menu to retrieved
          required: true
      responses:
        "200":
          description: Menu retrieved successfully
          schema:
            $ref: "#/definitions/MenuData"
        "401":
          description: Unauthorized
        "404":
          description: Menu not find
        "500":
          description: Interal Server Error
    put:
      security:
       - Bearer: []
      tags:
      - Menus Module
      summary: Update a specified menu
      description: Update a menu from the user with the menu id
      parameters:
        - in: path
          type: string
          name: menuId
          description: id of the menu to update
          required: true
        - in: body
          name: menu informations
          schema: 
            $ref: "#/definitions/MenuData"
          description: object with menu data
          required: true
      responses:
        "200":
          description: Menu update successfully
        "400":
          description: Bad request in body
        "401":
          description: Unauthorized
        "404":
          description: Menu not find
        "500":
          description: Interal Server Error
    delete:
      security:
       - Bearer: []
      tags:
      - Menus Module
      summary: Delete a specified menu
      description: delete a menu from the user with the menu id
      parameters:
        - in: path
          type: string
          name: menuId
          description: id of the menu to delete
          required: true
      responses:
        "200":
          description: Menu deleted successfully
        "401":
          description: Unauthorized
        "404":
          description: Menu not find
        "500":
          description: Interal Server Error
securityDefinitions:
  Bearer:
    type: apiKey
    in: header
    name: authToken
definitions:
  Login:
    properties:
      email:
        type: string
        description: user email
        example: example@gmail.com
  UserMenusResponse:
    type: array
    items:
      type: object
      properties:
        id:
          type: string
          example: "123456789"
        imgLink:
          type: string
        creationDate:
          type: string
  MenuData:
    type: object
    properties:
      toDefine: 
        type: string
      toDefine2: 
        type: string
  ApiResponse:
    type: object
    properties:
      status:
        type: integer
      message:
        type: string
      body:
        type: object
        properties:
          id:
            type: string
          email:
            type: string